Updated navbar + started hero
I updated the navbar with the team logo, and placing it next to the brand.
I also changed the background color of the navbar to match the logo background

Then I started the "hero" page process.
Should keep this simple & minimal text / or atleast not overwhlem with text

Also added footer