Fixed navbar + footer, added images
I fixed the navbar, specifically the unordered list items. Previously they were awkardly centered or right next to the brand logo. Now they're on the right side of the navbar.

The footer was also for some reason incldued in the a the "main" div/section block. Now it's separate, below it.

also added a few images of previous agmes; going to use them possiblty on the first page